restructures tests for json response #42
